requirements.txt
# run "pywin32_postinstall -install" command in last, win32con win32gui must be installed

django
djangorestframework
django-cors-headers
jdatetime
django-jalali
phonenumbers
django-phonenumber-field
undetected-chromedriver
selenium-stealth
opencv-python
opencv-python-headless
# scipy         # insalled by other packages
paddlepaddle
paddleocr
webdriver-manager
fake-useragent
pytz      # localize timezone to work celery ok
django-environ
djangorestframework-simplejwt
python-telegram-bot[socks]      # for telegram bot, socks required in linux
gunicorn
psutil
pywin32
channels
pywinauto
pygetwindow
daphne  # daphne -b 0.0.0.0 -p 8000 nobat.asgi:application
channels_postgres  # required for socket CHANNEL_LAYERS, if set on postgres
#psycopg[binary]     # used by channels
psycopg2-binary     # used for django (postgres backend), apscheduler
apscheduler
#pywin32_postinstall -install
